+++ src/config/mod.rs
@@ -1,2 +1,3 @@
 pub mod cameras;
 pub mod credentials;
+pub mod writer;